LPS-178068 Create ResourceClientConfiguration
