+++ workspaces/liferay-partner-workspace/client-extensions/liferay-partner-oauth-application-headless-server/src/main/java/com/liferay/partner/services/ObjectDefinitionService.java
@@ -17,7 +17,7 @@ package com.liferay.partner.services;
 import com.liferay.object.admin.rest.client.dto.v1_0.ObjectDefinition;
 import com.liferay.object.admin.rest.client.pagination.Page;
 import com.liferay.object.admin.rest.client.resource.v1_0.ObjectDefinitionResource;
-import com.liferay.partner.PartnerEnableWebSecurity.ResourceClientConfiguration;
+import com.liferay.partner.util.ResourceClientConfiguration;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
